# docs/specs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Stratum — organizations, accounts, billing, and notifications.",
        "title": "Stratum API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/invitations/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Accepts an invitation token for the caller's own (authenticated) email; the token must match the caller's email.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "Accept an invitation",
                "parameters": [
                    {
                        "description": "Invitation token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/acceptInvitationRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "422": {
                        "description": "invitation not found, expired, email mismatch, already a member, or plan limit reached",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
        "/invitations/decline": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Declines and deletes an invitation for the caller's own (authenticated) email; the original inviter is notified in-app.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "Decline an invitation",
                "parameters": [
                    {
                        "description": "Invitation token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/declineInvitationRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "422": {
                        "description": "invitation not found, expired, email mismatch, or already a member",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
        "/invitations/preview": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Read-only invitation details (organization name, role, inviter) for the caller's own email, before committing to accept.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "Preview an invitation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invitation token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Payload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/invitationPreview"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "422": {
                        "description": "missing token, not found, expired, email mismatch, or already a member",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
        "/invitations/request-new": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Notifies the original inviter that the caller's invitation (identified by an expired/invalid token) needs to be resent — does not resend automatically.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "Request a fresh invitation",
                "parameters": [
                    {
                        "description": "Expired/invalid invitation token",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requestNewInvitationRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "422": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
...
        "/me/invitations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns pending invitations sent to the caller's authenticated email, across every organization.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "List the caller's pending invitations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Payload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/myInvitationRecord"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
...
        "/organizations/{organizationID}/invitations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "List organization invitations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "organizationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Payload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/invitationRecord"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "403": {
                        "description": "admin role required",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organization"
                ],
                "summary": "Invite a member by email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "organizationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Invitation fields",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/createInvitationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Payload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/invitationRecord"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "403": {
                        "description": "admin role required",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "422": {
                        "description": "validation failed or invitee already a member",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
        "/organizations/{organizationID}/invitations/{invitationID}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "organization"
                ],
                "summary": "Revoke an invitation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization ID",
                        "name": "organizationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Invitation ID",
                        "name": "invitationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "403": {
                        "description": "admin role required",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Payload": {
            "type": "object",
            "properties": {
                "data": {},
                "next_cursor": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/pagination"
                },
                "status": {
                    "$ref": "#/definitions/Status"
                }
            }
        },
...
        "Status": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "acceptInvitationRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "createInvitationRequest": {
            "type": "object",
            "required": [
                "email",
                "role"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "admin",
                        "member"
                    ]
                }
            }
        },
...
        "declineInvitationRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "invitationPreview": {
            "type": "object",
            "properties": {
                "invitedByEmail": {
                    "type": "string"
                },
                "invitedEmail": {
                    "description": "InvitedEmail is only populated on an ErrInvitationEmailMismatch\nreturn — the address the invitation actually targets.",
                    "type": "string"
                },
                "organizationID": {
                    "type": "string"
                },
                "organizationName": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "invitationRecord": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invited_by": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "myInvitationRecord": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invited_by": {
                    "type": "string"
                },
                "invited_by_email": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "organization_name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "token": {
                    "description": "Token IS exposed here (unlike the admin-facing invitationRecord,\nwhich keeps json:\"-\") so the caller's own onboarding screen can\none-click POST /invitations/accept without a redirect — this\nendpoint is already scoped to the authenticated caller's own email,\nthe same audience the token was already emailed to.",
                    "type": "string"
                }
            }
        },
...
        "pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "requestNewInvitationRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token issued by Supabase auth.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}